replaced longs with uint32 references

+++ inc/ht_sched.hpp
@@ -39,10 +39,10 @@ namespace HT_SCHED
         HT_TASK::Task*                  _taskQueue[HT_SCHED_MAX_TASKS]; // ordered list of task object pointers
         int                             _numTasks;
         std::function<unsigned long()>  _microsFunction;
-        unsigned long                   _timeOfNextExec;        // time at which the next scheduled function must trigger
-        unsigned long                   _intervalExecTimer;     // accumulation of time spent executing scheduled tasks
-        unsigned long                   _idleExecTimer;         // accumulation of time spent executing idle tasks
-        unsigned long                   _schedulerExecTimer;    // accumulation of time spent in the scheduler
+        uint32_t                        _timeOfNextExec;        // time at which the next scheduled function must trigger
+        uint32_t                        _intervalExecTimer;     // accumulation of time spent executing scheduled tasks
+        uint32_t                        _idleExecTimer;         // accumulation of time spent executing idle tasks
+        uint32_t                        _schedulerExecTimer;    // accumulation of time spent in the scheduler
 
         // variables for SchedMon
         float periodicUtilization;  // CPU utilization of periodic tasks
